Refactor pureset interpreter tests
